+++ backend/app/operations/lesson_operations.py
@@ -12,9 +12,9 @@ def add_lesson(group_id, date, start_time, end_time, homework=None, room=None, t
         if isinstance(date, str):
             date = datetime.strptime(date, "%Y-%m-%d").date()
         if isinstance(start_time, str):
-            start_time = datetime.strptime(start_time, "%H:%M:%S").time()
+            start_time = datetime.strptime(start_time, "%H:%M").time()
         if isinstance(end_time, str):
-            end_time = datetime.strptime(end_time, "%H:%M:%S").time()
+            end_time = datetime.strptime(end_time, "%H:%M").time()
 
     except ValueError as e:
         print(f"Error in date or time conversion: {str(e)}")