+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2023-12-01
+### Added
+- New Student controls page
+- Delete UCGs when updating student level
+- Added delete event route
+- New Asso Controls page
+
+### Changed
+- Switched to HH:MM format for time inputs
+
+
 ## [3.1.0] - 2023-11-11
 ### Added
 - Register page